requirements.txt
google-adk
python-dotenv
browser-use
langchain-google-genai
litellm
